Moved UI types to nube-sdk-types
