+++ packages/types/src/index.ts
@@ -1,3 +1,4 @@
 export type * from "./main";
 export type * from "./domain";
 export type * from "./utility";
+export type * from "./ui";
