+++ packages/types/src/ui/col/col.ts
@@ -1,0 +1,7 @@
+import type { BoxProps } from "../box/box";
+
+export type ColProps = Omit<BoxProps, "direction">;
+
+export type Col = ColProps & {
+	type: "col";
+};
