+++ packages/types/src/ui/types.ts
@@ -18,3 +18,18 @@ export type TxtModifier =
 	| "lowercase"
 	| "uppercase"
 	| "capitalize";
+
+export type NubeComponent = {
+	type: string;
+	[key: string]: unknown;
+};
+
+export type UISlot =
+	| "before_main_content"
+	| "after_main_content"
+	| "before_line_items"
+	| "after_line_items";
+
+export type UIRender = Record<UISlot, NubeComponent>;
+export type UIValue = Record<string, string>;
+export type UI = UIRender & { values: UIValue };
