+++ packages/ui/src/components/box/box.ts
@@ -1,24 +1,6 @@
-import type { FlexContent, FlexItems, NubeComponent, Size } from "../types";
+import { Box, BoxProps } from "@tiendanube/nube-sdk-types";
 
-export type BoxProps = {
-	children?: NubeComponent[];
-	width?: Size;
-	height?: Size;
-	margin?: Size;
-	padding?: Size;
-	gap?: Size;
-	direction?: "row" | "col";
-	reverse?: boolean;
-	background?: string;
-	color?: string;
-	justifyContent?: FlexContent;
-	justifyItems?: FlexItems;
-	alignItems?: FlexItems;
-	alignContent?: FlexContent;
-	borderRadius?: Size;
-};
-
-export const box = (props: BoxProps): NubeComponent => ({
+export const box = (props: BoxProps): Box => ({
 	type: "box",
 	...props,
 });
