+++ packages/ui/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./components/types";
 export * from "./components/box/box";
 export * from "./components/col/col";
 export * from "./components/row/row";
